Bug 1146369 - Add the ability to control the ratio of previous / latest amis launched; r=rail
Adds a new argument "--latest-ami-percentage" which impacts the ratio of previous to current spot amis being launched.
e.g. if 50 is entered, and the script decides to launch 10 new amis in an environment where we're currently running 5 instances with the previous ami and 5 with the previous, spot requests will be made for 5 previous and 5 latest ami types as well: to keep the ratio at 50%