Bug 1210137: restrict outbound flows from AWS (setup testing)
The idea is to restrict flows to the internet using security groups;
flows to other mozilla destinations (10.0.0.0/8) are always filtered
at the destination.  This is based on the flows explicitly allowed to
datacenter hosts, per the firewall tests.

This patch sets up for testing these outbound flows by creating new
security groups to which a few hosts can be manually reassigned.