Bug 1286430 - Clone mozilla-unified to the pooled storage path
This matches what we now do for AMIs (see bug 1232442).

Most jobs are configured to use pooled repository storage. The
path to the shared repo is the SHA-1 of revision 0.

We clone from the mozilla-unified repo because it has most of
the commits from the Firefox repos (heads from central, aurora,
beta, release, esrs).